Refactor code with improvements
